Linux-native CMake port and core-only cleanup

+++ native/src/ReachVariantTool/game_variants/components/megalo/compiler/parsing/alias.cpp
@@ -8,7 +8,7 @@
 namespace {
    using namespace Megalo;
    //
-   template<typename T, int I> const OpcodeFuncToScriptMapping* _check_rel_alias_name_against_opcodes(const std::array<T, I>& list, const QString& name, const OpcodeArgTypeinfo* type) {
+   template<typename T, size_t I> const OpcodeFuncToScriptMapping* _check_rel_alias_name_against_opcodes(const std::array<T, I>& list, const QString& name, const OpcodeArgTypeinfo* type) {
       for (auto& function : list) {
          auto& mapping = function.mapping;
          if (mapping.arg_context < 0) // game namespace or no context
